requirements.txt
fastapi
uvicorn
jinja2
argon2-cffi
python-multipart
passlib[bcrypt]
python-jose[cryptography]